requirements.txt
altair==5.5.0
altair-data-server==0.4.1
altair-viewer==0.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
dash-bootstrap-components==2.0.3
Flask==3.0.3
fonttools==4.58.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.42.0
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
plotly==6.1.2
plotly-express==0.4.1
portpicker==1.6.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.25.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
speedtest-cli==2.1.3
statsmodels==0.14.4
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
wordcloud==1.9.4
zipp==3.23.0
